- removed output context; - added deletOnExit flag for dynamic and tmp contexts; - fixed conversion of audio (pcm instead of wav); - trim dynamic parameters values from conversion.xml; - DPP format builder: use input metadata.xml and audiomap.xml - DPP format builder: fill dynamic context correctly - supported creation of default audiomap.xml (with default mapping). - added tests.

+++ src/main/java/com/netflix/imfutility/AbstractFormatBuilder.java
@@ -2,24 +2,26 @@ package com.netflix.imfutility;
 
 import com.netflix.imfutility.asset.AssetMap;
 import com.netflix.imfutility.asset.AssetMapParser;
-import com.netflix.imfutility.config.ConfigProvider;
+import com.netflix.imfutility.config.ConfigXmlProvider;
 import com.netflix.imfutility.conversion.ConversionEngine;
 import com.netflix.imfutility.conversion.ConversionNotAllowedException;
-import com.netflix.imfutility.conversion.ConversionProvider;
+import com.netflix.imfutility.conversion.ConversionXmlProvider;
 import com.netflix.imfutility.conversion.SilentConversionChecker;
+import com.netflix.imfutility.conversion.templateParameter.context.CustomParameterValue;
 import com.netflix.imfutility.conversion.templateParameter.context.TemplateParameterContextProvider;
 import com.netflix.imfutility.cpl.CplContextBuilder;
 import com.netflix.imfutility.mediainfo.MediaInfoContextBuilder;
 import com.netflix.imfutility.mediainfo.MediaInfoException;
 import com.netflix.imfutility.xml.XmlParsingException;
 import com.netflix.imfutility.xsd.conversion.FormatConfigurationType;
-import com.netflix.imfutility.xsd.conversion.ParamType;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.net.URL;
 
 /**
  * The base class responsible for conversion to a destination format.
@@ -47,22 +49,31 @@ public abstract class AbstractFormatBuilder {
 
     protected final Format format;
     protected final String configXml;
-    protected final String conversionXml;
 
-    protected ConfigProvider configProvider;
-    protected ConversionProvider conversionProvider;
+    protected ConfigXmlProvider configProvider;
+    protected ConversionXmlProvider conversionProvider;
     protected FormatConfigurationType formatConfigurationType;
     protected TemplateParameterContextProvider contextProvider;
+
+    protected String conversionXml;
     protected String workingDir;
     protected AssetMap assetMap;
 
-    public AbstractFormatBuilder(Format format, String configXml, String conversionXml) {
-        this.format = format;
-        this.configXml = configXml;
-        this.conversionXml = conversionXml;
+    public AbstractFormatBuilder(Format format, String configXml) {
+        this(format, configXml, null);
     }
 
-    public final void build(String cplXml, String assetmapXml) {
+    public AbstractFormatBuilder(Format format, String configXml, String defaultWorkingDirectory) {
+        this.format = format;
+        this.configXml = configXml;
+        URL defaultConversionXmlUrl = ClassLoader.getSystemClassLoader().getResource("xml/conversion.xml");
+        if (defaultConversionXmlUrl != null) {
+            this.conversionXml = defaultConversionXmlUrl.getPath();
+        }
+        this.workingDir = defaultWorkingDirectory;
+    }
+
+    public final void build(String impFolder, String cplXml) {
         try {
             logger.info("Starting conversion to '{}' format\n", format.getName());
 
@@ -72,19 +83,17 @@ public abstract class AbstractFormatBuilder {
             // 2. clear working dir
             cleanWorkingDir();
 
-            // 3. create logs dir in the working fdir
+            // 3. create logs dir in the working dir
             createLogsDir();
 
             // 4. build IMF CPL contexts
-            buildCplContext(cplXml, assetmapXml);
+            buildCplContext(impFolder, cplXml);
 
-            // 5. fill dynamic and output contexts
+            // 5. fill dynamic context
             buildDynamicContext();
-            buildOutputContext();
 
             // 6. build Media Info contexts (get resource parameters such as channels_num, fps, sample_rate, etc.)
             buildMediaInfoContext();
-
 
             // 7. select a conversion config within format.
             selectConversionConfig();
@@ -97,9 +106,8 @@ public abstract class AbstractFormatBuilder {
             convert();
             postConvert();
 
-
             // 10. delete tmp files.
-            //deleteTmpFiles();
+            deleteTmpFiles();
 
             logger.info("Conversion to '{}' format: OK\n", format.getName());
         } catch (Exception e) {
@@ -108,20 +116,36 @@ public abstract class AbstractFormatBuilder {
 
     }
 
-    protected void init(String configXml, String conversionXml) throws XmlParsingException {
+    protected void init(String configXml, String conversionXml) throws XmlParsingException, FileNotFoundException {
         logger.info("Initializing...");
 
+        // 1. Reading and parsing config.xml
         logger.info("Reading config.xml: {}", configXml);
-        this.configProvider = new ConfigProvider(configXml);
+        this.configProvider = new ConfigXmlProvider(configXml);
         logger.info("Config.xml is processed: OK");
 
+        // 2. check for alternative conversion.xml
+        if (configProvider.getConfig().getConversionConfig() != null) {
+            this.conversionXml = configProvider.getConfig().getConversionConfig();
+            logger.info("Using alternative conversion.xml: {}", conversionXml);
+        }
+        if (this.conversionXml == null) {
+            throw new ConversionException("Conversion.xml is not found in neither default location nor config.xml");
+        }
+
+        // 3. Reading and parsing conversion.xml
         logger.info("Reading conversion.xml: {}", conversionXml);
-        this.conversionProvider = new ConversionProvider(conversionXml, format);
+        this.conversionProvider = new ConversionXmlProvider(conversionXml, format);
         logger.info("Conversion.xml is processed: OK");
 
+        // 4. setting working directory
         this.workingDir = configProvider.getConfig().getWorkingDirectory();
+        if (this.workingDir == null) {
+            throw new ConversionException("Working directory must be specified either in config.xml or as an input parameter");
+        }
         logger.info("Working directory: {}", this.workingDir);
 
+        // 5. Init the context provider
         this.contextProvider =
                 new TemplateParameterContextProvider(configProvider.getConfig(), conversionProvider.getFormat(), workingDir);
 
@@ -146,15 +170,22 @@ public abstract class AbstractFormatBuilder {
         logger.info("Created external tools logging directory: OK\n");
     }
 
-    private void buildCplContext(String cplXml, String assetmapXml) throws XmlParsingException {
+    private void buildCplContext(String impFolder, String cplXml) throws XmlParsingException, FileNotFoundException {
         logger.info("Building CPL contexts...");
 
-        logger.info("Parsing ASSETMAP.xml ('{}')...", assetmapXml);
-        this.assetMap = new AssetMapParser().parse(assetmapXml);
+        File impDir = new File(impFolder);
+        if (!impDir.isDirectory()) {
+            throw new FileNotFoundException(String.format("Invalid IMP directory: '%s' not found or not a directory", impDir.getAbsolutePath()));
+        }
+
+        File assetMapFile = new File(impFolder, Constants.ASSETMAP_FILE);
+        logger.info("Parsing ASSETMAP.xml ('{}')...", assetMapFile.getAbsolutePath());
+        this.assetMap = new AssetMapParser().parse(assetMapFile.getAbsolutePath());
         logger.info("Parsed ASSETMAP.xml: OK");
 
-        logger.info("Parsing CPL ('{}')...", cplXml);
-        new CplContextBuilder(contextProvider, assetMap).build(cplXml);
+        File cplFile = new File(impFolder, cplXml);
+        logger.info("Parsing CPL ('{}')...", cplFile.getAbsolutePath());
+        new CplContextBuilder(contextProvider, assetMap).build(cplFile.getAbsolutePath());
         logger.info("Parsed CPL: OK");
 
         logger.info("Built CPL contexts: OK");
@@ -167,14 +198,6 @@ public abstract class AbstractFormatBuilder {
     }
 
     protected abstract void doBuildDynamicContext();
-
-    private void buildOutputContext() {
-        logger.info("Building Output context...");
-        doBuildOutputContext();
-        logger.info("Built Output context: OK");
-    }
-
-    protected abstract void doBuildOutputContext();
 
     private void buildMediaInfoContext() throws XmlParsingException, IOException, MediaInfoException {
         logger.info("Building Metadata Info contexts...");
@@ -192,16 +215,14 @@ public abstract class AbstractFormatBuilder {
     }
 
     private void checkForSilentConversion() throws ConversionNotAllowedException {
-        logger.info("Checking whether it's allowed by config.xml to silently convert to destination parameters if they don't macth...");
+        logger.info("Checking whether it's allowed by config.xml to silently convert to destination parameters if they don't match...");
         new SilentConversionChecker(contextProvider, formatConfigurationType, configProvider).check();
         logger.info("Checked: silent conversion is either allowed or not needed.");
     }
 
-    protected void preConvert() throws IOException, XmlParsingException {
-    }
+    protected abstract void preConvert() throws IOException, XmlParsingException;
 
-    protected void postConvert() throws IOException, XmlParsingException {
-    }
+    protected abstract void postConvert() throws IOException, XmlParsingException;
 
     private void convert() throws IOException {
         logger.info("Starting conversion...");
@@ -215,11 +236,15 @@ public abstract class AbstractFormatBuilder {
         logger.info("Deleting tmp files created during conversion...");
 
         boolean success = true;
-        for (ParamType tmpParam : contextProvider.getTmpContext().getAllParameters()) {
-            success &= doDeleteTmpFile(tmpParam.getValue());
+        for (CustomParameterValue tmpParam : contextProvider.getTmpContext().getAllParameters()) {
+            if (tmpParam.isDeleteOnExit()) {
+                success &= doDeleteTmpFile(tmpParam.getValue());
+            }
         }
-        for (String paramValue : contextProvider.getDynamicContext().getAllParameters()) {
-            success &= doDeleteTmpFile(paramValue);
+        for (CustomParameterValue paramValue : contextProvider.getDynamicContext().getAllParameters()) {
+            if (paramValue.isDeleteOnExit()) {
+                success &= doDeleteTmpFile(paramValue.getValue());
+            }
         }
 
         if (success) {
